+++ src/Sources.hpp
@@ -6,7 +6,6 @@
 
 // Custom headers
 #include "GfxFileWriters.h"
-#include "Config.h"
 
 // Add your plugin files here
 #include "Screenshot.hpp"