+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-Initial release
+- Replaced saving plugin options in custom Screenshot.ini with regular Gothic.ini usage (Screenshot section)